Move GRACE map components to GRACE module.
