+++ src/app/shared/modules/grace/grace.module.ts
@@ -6,6 +6,7 @@ import { GraceService } from './grace.service';
 import { GraceStyleComponent } from './grace-style.component';
 import { GraceStyleService } from './grace-style.service';
 import { GraceGraphModalComponent } from './grace-graph.modal.component';
+import { OlMapGraceDataComponent } from './olmap.grace-data.component';
 import { StyleChooserModalComponent } from './style-chooser.modal.component';
 
 import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
@@ -27,8 +28,8 @@ PlotlyViaCDNModule.setPlotlyBundle("basic");
     PlotlyViaCDNModule,
     ColorPickerModule
   ],
-  declarations: [ GraceGraphModalComponent, StyleChooserModalComponent, GraceStyleComponent ],
-  entryComponents: [ GraceGraphModalComponent, StyleChooserModalComponent, GraceStyleComponent ],
+  declarations: [ GraceGraphModalComponent, StyleChooserModalComponent, GraceStyleComponent, OlMapGraceDataComponent ],
+  entryComponents: [ GraceGraphModalComponent, StyleChooserModalComponent, GraceStyleComponent, OlMapGraceDataComponent ],
   providers: [ GraceService, GraceStyleService ]
 })
 export class GraceModule { }
